% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_report.R
\name{cast.and.fill}
\alias{cast.and.fill}
\title{cast.and.fill}
\usage{
cast.and.fill(data, condition.var, time.var, measure.var, na.fill)
}
\arguments{
\item{data}{A data table in long format.}

\item{condition.var}{Character vector, names of variables used for rows in
casting, define the conditions of the experiment.}

\item{time.var}{Character, name of Time variable used for columns in casting}

\item{measure.var}{Character, name of measurement variable, on which PCA is
to be performed.}

\item{na.fill}{Numeric, values to replace NAs}
}
\value{
Two matrices. "mat" is a wide character matrix which contains the
  casted "condition.var" as first columns followed by casted measurements.
  "mat2" is a numeric wide matrix, which is essentially the same with
  conditions trimmed.
}
\description{
Cast a data table from long to wide and fill missing values. Output is
suitable for PCA.
}
