% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circular.cc}
\alias{circular.cc}
\title{circular.cc}
\usage{
circular.cc(x, y)
}
\arguments{
\item{x, y}{numerical vector.}
}
\value{
The vector of cross-correlation between x and y at all possible lags.
}
\description{
Cross-correlation with circular boundaries.
}
\examples{
# Same pattern shifted by 2 time units
x <- rep(c(0,1,2), 10)
y <- rep(c(2,0,1), 10)
xy.cc <- circular.cc(x, y)
plot(names(xy.cc),xy.cc, type="h",
xlab = "Lag", ylab = "Correlation")
abline(h = 0, lty = "dashed", col = "blue")

}
