% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classical_decomposition.R
\name{classical.decomposition}
\alias{classical.decomposition}
\alias{plot_decomposition}
\title{Decompose a time series into trend, seasonal and remainder component}
\usage{
classical.decomposition(ts, period, robust = T)

plot_decomposition(decomposition, main = "Time series decomposition")
}
\arguments{
\item{ts:}{a time series or numerical vector}

\item{period:}{number of time points in one cycle}

\item{robust:}{if T, the median of all points of a cycle are used to
determined seasonal component; if F, the mean is used instead.}
}
\value{
a 3xn numeric matrix, with column trend, season and remainder
}
\description{
The model is assumed to be additive, i.e. the time series is of the form:
y(t) = trend(t) + seasonal(t) + remainder(t)
}
\details{
Trends is determined using a rolling mean with window width corresponding to
the signal period, extremities are padded with linear interpolation from the
first and last 2 measures.

Season is determined by summarizing points "period-wise". Specifically,
let's consider a signal, which has 5 periods, each composed of 15 points.
The 1st point of a season is determined by taking the mean (or median if
robust=TRUE) of the 1st point of each of the 5 periods. Once all 15 points
have been summarized, this 'unit season' is repeated 5 times to form the
seasonal component.
}
\section{Functions}{
\itemize{
\item \code{plot_decomposition}: utility plot function
}}

\note{
Good explanation of the decomposition at https://www.otexts.org/fpp/6/3
}
\examples{
# Regular motif of length 15 repeated 5 times + Linear Trend
x <- rep(1:15, 5)
x <- x + seq(0, 10, length.out = length(x))
x_decomp <- classical.decomposition(ts = x, period = 15)
plot_decomposition(x_decomp)
}
\seealso{
decompose
}
