% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{complete.time.series}
\alias{complete.time.series}
\title{complete.time.series}
\usage{
complete.time.series(data, cond.col, lab.col, time.col, time.vector, meas.col)
}
\arguments{
\item{data}{a data.table in long format with at least 4 columns:
condition, label, time and measurement.}

\item{cond.col}{column name for grouping. Typically an ID for experimental
conditions.}

\item{lab.col}{column name for second grouping. Typically an ID for
trajectories. This ID can be shared between different conditions (first
grouping).}

\item{time.col}{column name of time.}

\item{time.vector}{numerical vector. Over which time should ALL time series
span? Missing times will be added to trajectory where it's not present.}

\item{meas.col}{column name with measurements.}
}
\value{
A data.table with extra rows for missing measurements.
}
\description{
Add rows for missing measurements.
}
\examples{
# Simulate 10 phase-shifted sinusoids with 3 different level of noises
# ("experimental conditions", first grouping).
x <- multi_sims(type = "ps", noises = c(0.2,0.4), n = 10)
plot_sim(x)
# Remove 50 random rows
x[, row.nber := 1:nrow(x)] # For manually checking missing values
row_to_del <- sample(1:nrow(x), size = 50, replace = F)
row_to_keep <- setdiff(1:nrow(x), row_to_del)
x <- x[row_to_keep]
# Recreate the missing rows
x_complete <- complete.time.series(data = x,
 cond.col = "noise", lab.col = "variable",
 time.col = "Time", time.vector = unique(x$Time),
 meas.col = "value")

}
