% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatAllFeat}
\alias{FeatAllFeat}
\title{FeatAllFeat}
\usage{
FeatAllFeat(y, basal, start.lag.grow, end.exp.dec, ...)
}
\arguments{
\item{y}{a numerical vector from which features are extracted.}

\item{basal}{a normalization constant to bring base of the peak around 0.}

\item{start.lag.grow}{Index of y where ascending phase starts. See
FeatLagGrow, argument 'start'.}

\item{end.exp.dec}{Index of y where descending phase ends. See FeatExpDec,
argument 'end'.}

\item{...}{Additional arguments for FeatFWHM.}
}
\value{
A list of 13 features: \itemize{ \item mini: minimum of y \item maxi:
  maximum of y \item diff.min.max: difference maxi-mini \item time.min: index
  of y where it is minimized \item time.max: index of y where it is maximized
  \item max.amp: amplitude of peak. Obtained by subtracting constant basal.
  See FeatMaxAmplitude. \item FWHM: Full-Width at Half maximum. Difference
  right-left. See FeatFWHM. \item left: time where half maximum of the peak
  is reached on its left flank. \item right: time where half maximum of the
  peak is reached on its right flank. \item grow.half,grow.lag: growth rate
  estimated by the slope of a linear regression of the ascending phase of the
  peak. 2 different methods are used to isolate this phase. First, it is
  defined as the time between which the signal has reached half of its
  maximum (left) and its maximum; see FeatHalfMaxGrow. Second, as the time
  between a hard-encoded starting time point (start.lag.grow) and peak
  maximum; see FeatLagGrow. \item dec.half,dec.exp: decay rate estimated by
  the slope of a linear regression of the descending phase of the peak. 2
  different methods are used to isolate this phase. First, it is defined as
  the time between which the signal has reached its maximum and half of its
  maximum (right); see FeatHalfMaxDec. Second, as the time between peak
  maximum and a hard-encoded endtime point (start.lag.grow); see FeatExpDec.
  In this case the signal is modelled by EXPONENTIAL DECAY. }
}
\description{
Extract all single-peak features
}
\examples{
signal <- dnorm(x = seq(-4,4, length.out = 500), mean = 0, sd = 1)
feats <- FeatAllFeat(y = signal, basal = 0, start.lag.grow = 100, end.exp.dec = 400)
plot(signal)
# Half-maximum points
abline(h = max(signal)/2, col = "green")
abline(v = c(feats$left, feats$right), col = c("blue", "red"))

# Growth and Decay slopes according to half-max methods
growth0 = c(feats$left, max(signal)/2)
growth1 = c(which.max(signal), max(signal)/2 + feats$grow.half * (which.max(signal)-feats$left))
segments(x0 = growth0[1], y0 = growth0[2], x1 = growth1[1], y1 = growth1[2], col = "blue", lwd = 4, lty = "dashed")

decay0 = c(which.max(signal), max(signal))
decay1 = c(feats$right, max(signal) + feats$dec.half * (feats$right-which.max(signal)))
segments(x0 = growth0[1], y0 = growth0[2], x1 = growth1[1], y1 = growth1[2], col = "blue", lwd = 4, lty = "dashed")

}
