% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatDiffMinMax}
\alias{FeatDiffMinMax}
\title{FeatDiffMinMax}
\usage{
FeatDiffMinMax(y)
}
\arguments{
\item{y}{a numerical vector}
}
\value{
List of 5:
  "min": minimum of y
  
  "max": maximum of y
  
  "diff.min.max": $max - $min
  
  "time.min": first index at which $min is reached in y
  
  "time.max": first index at which $max is reached in y
}
\description{
Difference between minimum and maximum value.
}
