% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatExpDec}
\alias{FeatExpDec}
\title{FeatExpDec}
\usage{
FeatExpDec(y, end)
}
\arguments{
\item{y}{numeric, typically measured variable!}

\item{end}{int, at which index should the regression end.}
}
\value{
slope of the linear regression, estimator of the decay rate.
}
\description{
Estimate decay rate of the peak. Fit a linear model on the descending phase,
using exponential decay model. The fit is performed betwenn peak maximum and
til the index specified by end and is enforced to pass through peak maximum.
}
\seealso{
FeatHalfMaxDec
}
