% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatHalfMaxGrow}
\alias{FeatHalfMaxGrow}
\title{FeatHalfMaxGrow}
\usage{
FeatHalfMaxGrow(y, ...)
}
\arguments{
\item{y}{numeric, typically measured variable.}

\item{...}{extra arguments for FeatFWHM. Notably 'basal' for setting base of
the peak around 0.}
}
\value{
Slope of the linear regression, estimator of the growth rate.
}
\description{
Estimate growth rate of the peak. Fit a linear model on growing phase of a
peak between (possibly interpolated) time at which half-max of the peak is
reached and maximum of the peak
}
\seealso{
FeatFWHM, FeatLagGrow
}
