% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.period}
\alias{get.period}
\title{get.period}
\usage{
get.period(x)
}
\arguments{
\item{x}{a numerical vector.}
}
\value{
Period (1/frequency) of maximum power density
}
\description{
Compute power spectrum and returns period (1/frequency) at which
its density is maximized.
}
\examples{
# Regular motif of length 15 repeated 5 times + Linear Trend
x <- rep(1:15, 5)
x <- x + seq(0, 10, length.out = length(x))
x_period <- get.period(x)
# Use the period to perform classical decomposition
x_decomp <- classical.decomposition(ts = x, period = x_period)
plot_decomposition(x_decomp)

}
