% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchrony.R
\name{max_cc}
\alias{max_cc}
\title{max.cc}
\usage{
max_cc(x, y, use.circular = TRUE, plot = FALSE, ...)
}
\arguments{
\item{x, }{y numerical vector}

\item{use.circular}{logical, whether to use circular cross-correlation.}

\item{plot}{logical. If TRUE, plots the cross-correlation histogram.}

\item{...}{additional parameters for the cross-correlation if use.circular is FALSE.
Of particular interest is lag.max. See ?ccf.}
}
\value{
A list of 2. $corr contains the maximum of correlation and
$lag the lag at which it was reached.
}
\description{
Get maximal cross-correlation between 2 numerical vectors. Only 1st position
is returned.
}
\examples{
# 2 identical signals, shifted by a lag 3
pattern <- c(0.25, 0.5, 0.75, 1)
x <- rep(pattern, 10)
y <- c(x[4:length(x)],x[1:3])
par(mfrow=c(3,1))
plot(x, type = "h", main = "x", ylim = c(-0.2,1.2))
plot(y, type = "h", main = "y", ylim = c(-0.2,1.2))
# With circular cc, correlation is strictly 1
max_cc(x, y, use.circular = T, plot = T)

}
