% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_peak_features.R
\name{MPFeatAmp_euclidmean}
\alias{MPFeatAmp_euclidmean}
\title{MPFeatAmp_euclidmean}
\usage{
MPFeatAmp_euclidmean(x, window.size)
}
\arguments{
\item{x}{a numerical vector.}

\item{window.size}{integer, width of window for rolling mean.}
}
\value{
The Euclidean distance between x and its rolling mean.
}
\description{
Compute Euclidean distance between a signal and its rolling mean.
Rolling mean is extended at extremeties by linear extrapolation, see ?rollex.
Aims at giving an estimation of peak amplitude in oscillating signal.
}
\seealso{
MPFeatAmp_seasonal, MPFeatAmp_variation
}
