% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_peak_features.R
\name{MPFeatAmp_seasonal}
\alias{MPFeatAmp_seasonal}
\title{MPFeatAmp_seasonal}
\usage{
MPFeatAmp_seasonal(x, window.size, robust = TRUE)
}
\arguments{
\item{x}{a numerical vector.}

\item{window.size}{integer, width of window for rolling mean used to extract
the trend component. See ?classical.decomposition.}

\item{robust}{logical. If TRUE, seasonal component is estimated by taking
the median of values at each stage of a season. If FALSE, uses mean.
See ?classical.decomposition.}
}
\value{
The difference between maximum and minimum of the seasonal component.
}
\description{
Decompose signal with classical decomposition and returns amplitude of the
seasonal component.
Aims at giving an estimation of peak amplitude in oscillating signal.
}
\seealso{
MPFeatAmp_euclidmean, MPFeatAmp_variation
}
