% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_peak_features.R
\name{MPFeatAmp_variation}
\alias{MPFeatAmp_variation}
\title{MPFeatAmp_variation}
\usage{
MPFeatAmp_variation(x, window.size, trim = NULL, robust = TRUE)
}
\arguments{
\item{x}{a numerical vector.}

\item{window.size}{integer, width of window for rolling mean used to extract
the trend component. See ?classical.decomposition.}

\item{trim}{optional. Gives the minimum absolute value of remainder to be
considered.}

\item{robust}{logical. If TRUE, seasonal component is estimated by taking
the median of values at each stage of a season. If FALSE, uses mean.
See ?classical.decomposition.}
}
\value{
The mean (trimmed) absolute remainder.
}
\description{
Decompose signal with classical decomposition and returns mean of the
(trimmed) absolute remainder component.
Aims at at giving an estimation of the variation of peak amplitudes
in oscillating signal.
}
\seealso{
MPFeatAmp_euclidmean, MPFeatAmp_seasonal
}
