% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_peak_features.R
\name{MPFeatTrend_extrema}
\alias{MPFeatTrend_extrema}
\title{MPFeatTrend_extrema}
\usage{
MPFeatTrend_extrema(x, window.size, what, robust = FALSE)
}
\arguments{
\item{x}{a numerical vector.}

\item{window.size}{integer, width of window for maxima detection.
See ?detect.peak.}

\item{what}{character, one of "minima" or "maxima". Define whether regression
should be performed on maxima or minima of peaks.}

\item{robust}{logical. If TRUE perform robusts linear regression.
Median-Based Linear Models are used, see ?mblm::mblm}
}
\value{
A list of 4.
\itemize{
 \item $trend, slope of the linear model
 \item $model, the full linear model
 \item $extremes, x extrema values and their index
 \item $type, type of extrema, "mini" or "maxi"
 }
}
\description{
Isolate local extrema and perform a linear regression.
Aims at giving an estimation of long-term trend in multi-peak signals.
}
\examples{
x <- sim_phase_shifted_with_fixed_trend(n = 1, noise = 0, slope = 0.1)
x.trend <- MPFeatTrend_extrema(x = x$value, window.size = 5, what = "maxi")
plot(x$value, type = "b")
abline(x.trend$model, col = "blue", lwd = 2, lty = "dashed")

}
\seealso{
MPFeatTrend_rollmean
}
