% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_peak_features.R
\name{MPFeatTrend_rollmean}
\alias{MPFeatTrend_rollmean}
\title{MPFeatTrend_rollmean}
\usage{
MPFeatTrend_rollmean(x, window.size, robust = FALSE)
}
\arguments{
\item{x}{a numerical vector.}

\item{window.size}{integer, width of window for rolling mean.}

\item{robust}{logical. If TRUE perform robusts linear regression.
Median-Based Linear Models are used, see ?mblm::mblm}
}
\value{
A list of 2: $model, the full linear model;
$trend, slope of the linear model.
}
\description{
Smooth by rolling mean and perform linear regression.
Rolling mean is extended at extremeties by linear extrapolation, see ?rollex.
Aims at giving an estimation of long-term trend in multi-peak signals.
}
\examples{
width <- 30
x <- sim_phase_shifted_with_fixed_trend(n = 1, noise = 0, slope = 0.1)
# Rolling mean extended by linear extrapolation.
# This is also performed in MPFeatTrend_rollmean.
x.roll <- rollex(x$value, width)
plot(x$value, type = "b")
lines(x.roll, col ="red", lwd = 2, lty = "dashed")
x.trend <- MPFeatTrend_rollmean(x = x$value, window.size = width)
abline(x.trend$model, col = "blue", lwd = 2, lty = "dashed")

}
\seealso{
MPFeatTrend_extrema
}
