% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{multi_sims}
\alias{multi_sims}
\title{multi_sims}
\usage{
multi_sims(type, noises, n, freq = 0.2, end = 50, ...)
}
\arguments{
\item{type}{character indicating which type of time series are to be
generated. Possible values are: \itemize{ \item ps: Phase-Shifted plain
sinusoids. \item pst: Phase-Shifted plain sinusoids with linear trend over
time. \item psd: Phase-Shifted and Damped sinusoids (chirp). \item na:
Noisy Amplitude, plain sinusoids with Gaussian white noise. \item nad:
Noisy Amplitude Damped, plain sinusoids with Gaussian white noise. \item
edls: Exponential Decay Lag Shifted. }}

\item{noises}{numerical vector giving levels of noise to generate the data.
Do not pass 0 (no noise) as this is already generated by default. The
feature controlled by this argument varies across types of simulations.
\itemize{ \item Phase-Shifted simulations ('psX'): standard deviation of
the Gaussian distribution used to shift the curves. \item Noisy-Amplitude
simulations ('naX'): standard deviation of the Gaussian distribution used
to generate amplitude noise at each time point. \item Exponential decay Lag
Shifted: standard deviation of the Gaussian distribution used to generate
the lag between excitation time and response time. } Typical values ranges
from 0 to 2.}

\item{n}{number of trajectories per noise level.}

\item{freq}{numeric, spacing between 2 time points.}

\item{end}{numeric, end time of simulations.}

\item{...}{Additional arguments specific to simulation type. See
corresponding sim_X functions.}
}
\value{
a data.table containing the trajectories in long format in 4 columns.
  "variable" indicates the ID of the trajectory. IDs are named V1, V2, ...,
  Vn for each level of noise.
}
\description{
Generate artificial oscillating time series of various types.
}
\details{
See ?sim_phase_shifted.
}
\note{
TODO - function is not optimized: it builds the simulation by growing a
  data.table. plot_sim can be used to visualize the output.
}
\examples{
x <- multi_sims(type = "ps", noises = seq(0.5, 2, 0.5), n = 10, freq = 0.5, end = 30)
plot_sim(x)

}
\seealso{
plot_sim, sim_phase_shifted, sim_phase_shifted_damped,
  sim_phase_shifted_with_fixed_trend, sim_noisy_amplitude,
  sim_noisy_amplitude_damped, sim_expodecay_lagged_stim
}
