% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNorm.R
\name{myNorm}
\alias{myNorm}
\title{myNorm}
\usage{
myNorm(in.dt, in.meas.col, in.rt.col = "RealTime", in.rt.min = 10,
  in.rt.max = 20, in.by.cols = NULL, in.robust = TRUE,
  in.type = "z.score")
}
\arguments{
\item{in.dt}{data.table, contain a column for: time, measurement(, grouping
columns like object label)}

\item{in.meas.col}{character, name of column to be normalised.}

\item{in.rt.col}{character, name of the column with time (or index)}

\item{in.rt.min}{numeric, start time of normalization period.}

\item{in.rt.max}{numeric, end time of normalization period.}

\item{in.by.cols}{character vector with 'by' columns to calculate
normalisation per group if NULL, no grouping is done}

\item{in.robust}{logical, whether robust measures should be used (median
instead of mean, mad instead of sd).}

\item{in.type}{type of normalization: 'z.score' or 'mean' (fi.e. old change
w.r.t. mean)}
}
\value{
A data.table with an additional column with normalized quantity. The
  name of additional column is the same as in.meas.col but with ".norm"
  suffix added.
}
\description{
Returns original dt with an additional column with normalized quantity.
Normalisation is based on part of the trajectory.
}
