% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_estimation.R
\name{noise_cv}
\alias{noise_cv}
\title{noise_cv}
\usage{
noise_cv(x)
}
\arguments{
\item{x}{a numerical matrix. Individuals in rows, measurements in columns.}
}
\value{
A vector with coefficient of variations of each column.
}
\description{
Coefficient of variation of each column of a matrix. Can be used to
get a noise estimation at each time point in an ensemble of time series.
}
\examples{
x <- rnorm(20)
x <- matrix(x, nrow = 4, ncol=5)
noise_cv(x)

}
