% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_report.R
\name{PCA.DB}
\alias{PCA.DB}
\title{PCA.DB}
\usage{
PCA.DB(pca, PC = NULL, PCvar = NULL, labels.pca, label.cluster)
}
\arguments{
\item{pca}{A pca object.}

\item{PC}{Which PCs should be used to compute the DB index. Consider only a
few components to use only leading trends in data.}

\item{PCvar}{If provided, replace PC. Instead of using a given set of PC,
will use as many as necessary to reach PCvar% of explained variance.}

\item{labels.pca}{A data.table or character matrix with a number of rows
equal to the number of time series used in pca. Each column contains a
different label for the associated trajectory. Note that column names must}

\item{label.cluster}{be provided.}
}
\value{
A numeric, the Davies-Bouldin index of the clustering.
}
\description{
Compute the Davies-Bouldin index for a pca clustering. This is used to
quantify how much groups are taken away in PCA space.
}
