% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{plot_sim}
\alias{plot_sim}
\title{plot_sim}
\usage{
plot_sim(data, x = "Time", y = "value", group = "variable",
  use.facet = T, facet = "noise", alpha = 0.2, plot = T)
}
\arguments{
\item{data}{data.table in long format}

\item{x}{column name of x axis (typically time)}

\item{y}{column name of y axis (typically measurements)}

\item{group}{column name of grouping (typically ID of a trajectory)}

\item{use.facet}{logical, if TRUE use facet to give a second grouping.}

\item{facet}{column name of second grouping for facetting}

\item{alpha}{numeric, set transparency of curves}

\item{plot}{logical, whether to plot the output}
}
\value{
A ggplot object.
}
\description{
Visualize a long data.table such as the one generated by multi_sims.
}
\examples{
x <- multi_sims(type = "ps", noises = seq(0.5, 2, 0.5), n = 10, freq = 0.5, end = 30)
plot_sim(x)

}
