% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_random_subset}
\alias{select_random_subset}
\title{Select Random}
\usage{
select_random_subset(in.file, ntraj, col.uniqID, col.time, min.obs = NULL,
  col.whatokeep = NULL, out.file = NULL)
}
\arguments{
\item{in.file}{A character or a variable for the current environment. If a
character, a path to a .csv file is expected.}

\item{ntraj}{Numeric. How many unique IDs to extract?}

\item{col.uniqID}{Character. Name of the column with unique IDs.}

\item{col.time}{Character. Name of the column with time.}

\item{min.obs}{Optional numeric. Minimum number of observations of an ID.}

\item{col.whatokeep}{Optional character. Name of the columns to be returned
in the output. If NULL returns all columns. Columns with uniqID and time
are always returned.}

\item{out.file}{Optional character. If provided the output is saved as a .csv
file to the path indicated by out.file. Otherwise, the new dataset is
returned.}
}
\value{
A data.table with the columns indicated in col.whatokeep.
}
\description{
Select a random subset of time series from a dataset by using unique ID
subset. Optionally, a minimum number of observation for the individual IDs
can be provided.
}
\examples{
# Simulate 10 phase-shifted sinusoids with 3 different level of noises
# ("experimental conditions", first grouping).
x <- multi_sims(type = "ps", noises = c(0.2,0.4), n = 10)
plot_sim(x)

# Read/export 5 random TS from/in environment
x[, uniqID := paste(noise, variable, sep ="_")]
x_subset <- select_random_subset(in.file=x, ntraj=5, col.uniqID="uniqID", col.time="Time")
}
