% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separability.R
\name{separability.measures}
\alias{separability.measures}
\title{separability.measures}
\usage{
separability.measures(Vector.1, Vector.2)
}
\arguments{
\item{Vector.1, Vector.2}{numerical vectors}
}
\value{
A vector of 5 measures: \itemize{ \item "jm": Jeffries-Matusita index.
  Bounded version of Bhattacharrya on (0,2). 0: identical dsitributions; 2:
  perfectly separable distributions. \item "bh": Bhattacharrya index Vary on
  (0, +inf). 0: identical distributions. \item "div": Kullback-Leibler
  Divergence. \item "tdiv": Transformed Divergence }
}
\description{
Compute a set of measures for separability of 2 distributions. from:
https://stats.stackexchange.com/a/78855/43610
}
