% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_peak_features.R
\name{FeatDiffMinMax}
\alias{FeatDiffMinMax}
\title{FeatDiffMinMax}
\usage{
FeatDiffMinMax(y)
}
\arguments{
\item{y}{a numerical vector}
}
\value{
List of 5:
\itemize{
  \item min: minimum of y
  \item max: maximum of y
  \item diff.min.max: $max - $min
  \item time.min: first index at which $min is reached in y
  \item time.max: first index at which $max is reached in y
}
}
\description{
Difference between minimum and maximum value.
}
